fix(VTreeview): unable to unselect child nodes in certain scenarios
this is a breaking change. parent nodes are no longer included in the selection.
instead two different selection types are offered, leaf and independent. leaf is
similar to existing behaviour except as mentioned parents are not included.
independent selects nodes independent of their position in tree hierarchy, and
allows for selection of parent nodes. selection does not propagate up or down.

solves #6903